# swagger_server/models/repo.py
# coding: utf-8

from __future__ import absolute_import
from .base_model_ import Model
from datetime import date, datetime
from typing import List, Dict
from ..util import deserialize_model


class Repo(Model):
    """
    NOTE: This class is auto generated by the swagger code generator program.
    Do not edit the class manually.
    """
    def __init__(self, name: str=None):
        """
        Repo - a model defined in Swagger

        :param name: The name of this Repo.
        :type name: str
        """
        self.swagger_types = {
            'name': str
        }

        self.attribute_map = {
            'name': 'name'
        }

        self._name = name

    @classmethod
    def from_dict(cls, dikt) -> 'Repo':
        """
        Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The Repo of this Repo.
        :rtype: Repo
        """
        return deserialize_model(dikt, cls)

    @property
    def name(self) -> str:
        """
        Gets the name of this Repo.

        :return: The name of this Repo.
        :rtype: str
        """
        return self._name

    @name.setter
    def name(self, name: str):
        """
        Sets the name of this Repo.

        :param name: The name of this Repo.
        :type name: str
        """

        self._name = name